chore: initial cleanup of code and moving to correct package
